requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
ipykernel
kagglehub
tensorflow
python-dotenv
scipy
PyYAML
pyspark
mlflow
catboost
xgboost
fastapi
uvicorn
boto3
opencv-python